[VIDEO] av not sync

+++ dtutils/dt_time.c
@@ -3,11 +3,11 @@
 #include <unistd.h>
 #include <stddef.h>
 
-uint64_t dt_gettime (void)
+int64_t dt_gettime (void)
 {
     struct timeval tv;
     gettimeofday (&tv, NULL);
-    return (uint64_t) (tv.tv_sec * 1000000 + tv.tv_usec);
+    return ((int64_t)(tv.tv_sec) * 1000000 + (int64_t)(tv.tv_usec));
 }
 
 int dt_usleep (unsigned usec)
